clang: Make libclang_frontend depend on libbcc
libclang does use functions from libbcc so it does depend on it.
When defining new symbols in bpf_common.cc, linker errors appear
when the same symbols are used in libclang. This is because of
incorrect linker dependency, this patch fixes the issue by making
sure the dependency is correctly tracked.

Signed-off-by: root <joelaf@google.com>