bcc/remote: Add support for bcc remotes
Add support for bcc remotes, this allows one to use bcc across systems.
In such a configuration, bcc would run on a different system than the
target running the bpf program. This allows us to have kernel sources
and the LLVM stack on a single machine such as a development machine
without having to setup all these on the target. It also allows us to
run clang on a different architecture than the target's architecture
thus facilitating cross compilation development.

Signed-off-by: Joel Fernandes <joelaf@google.com>