clang: Add support to build eBPF for user specified ARCH
BCC at the moment builds eBPF only considering the local architecture
instead of the one that the user's target kernel is running on.
For cross-compiler environments, the ARCH environment variable is
used to specify which ARCH to build the kernel for. In this patch we
add support to read either of 2 environment variables, ARCH and if
that's not set, then BCC_ARCH. Based on this we build the eBPF program
for the architecture specified.

This patch borrows some code from a patch doing a similar thing for
eBPF samples in the kenrel that I submitted recently [1]

[1] https://patchwork.kernel.org/patch/9961801/

Signed-off-by: Joel Fernandes <joelaf@google.com>